pyproject.toml
[tool.poetry]
name = "secuserve-liveness-module"
version = "0.1.0"
description = "A module for secuserve to reconize the liveness of faces"
authors = ["NicholasBlackburn1 <nickblackburn02@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
tensorflow = "2.3.1"
zmq = "^0.0.0"
pytest = "^6.2.5"
colorama = "^0.4.4"
opencv-python = "4.4.0.46"
torch = "1.9.0"
Cython = "0.29.24"
tqdm = "4.61.0"
matplotlib = "^3.4.3"
PyYAML = "^6.0"
scipy = "^1.7.1"
torchvision = "0.10.0"
onnxruntime = "^1.9.0"
keras = "^2.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
